[#2213] Dispatch auth logic depending on authProvider [client]
